+++ src/Validation/dtos/validation.requiredwithkeys.dto.ts
@@ -1,0 +1,7 @@
+import { ValidationGeneralDto } from "./validation.general.dto";
+
+export interface IValidationRequiredWithKeysDto extends Omit<ValidationGeneralDto, 'params' | 'customFunction'> {
+    params: {
+        keys: string[]
+    }
+}